+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientExternalAuthorizationServerTest.kt
@@ -383,10 +383,10 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
             val selectedCredential = credentialIdentifierInfos
                 .first { it.supportedCredentialFormat.format == CredentialFormatEnum.MSO_MDOC }
 
-            val offer = externalAuthorizationServer.credentialOfferWithPreAuthnForUser(
+            val offer = externalAuthorizationServer.offerWithPreAuthnForUserForSchemes(
                 user = dummyUser(),
                 credentialIssuer = credentialIssuer.metadata.credentialIssuer,
-                credentials = setOf(EuPidScheme to ISO_MDOC),
+                schemes = setOf(EuPidScheme to ISO_MDOC),
             )
             client.loadCredentialWithOfferReturningResult(offer, selectedCredential, null).getOrThrow().also {
                 it.shouldBeInstanceOf<CredentialIssuanceResult.Success>().also {