+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientTest.kt
@@ -305,10 +305,10 @@ val OpenId4VciClientTest by testSuite {
             val selectedCredential = credentialIdentifierInfos
                 .first { it.supportedCredentialFormat.format == CredentialFormatEnum.MSO_MDOC }
 
-            val offer = authorizationService.credentialOfferWithPreAuthnForUser(
+            val offer = authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = dummyUser(),
                 credentialIssuer = credentialIssuer.metadata.credentialIssuer,
-                credentials = setOf(EuPidScheme to ISO_MDOC),
+                schemes = setOf(EuPidScheme to ISO_MDOC),
             )
             val issuedCredential = client.loadCredentialWithOfferReturningResult(offer, selectedCredential, null)
                 .getOrThrow()
@@ -348,10 +348,10 @@ val OpenId4VciClientTest by testSuite {
             val selectedCredential = credentialIdentifierInfos
                 .first { it.supportedCredentialFormat.format == CredentialFormatEnum.MSO_MDOC }
 
-            val offer = authorizationService.credentialOfferWithPreAuthnForUser(
+            val offer = authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = dummyUser(),
                 credentialIssuer = credentialIssuer.metadata.credentialIssuer,
-                credentials = setOf(EuPidScheme to ISO_MDOC),
+                schemes = setOf(EuPidScheme to ISO_MDOC),
             )
             client.loadCredentialWithOfferReturningResult(offer, selectedCredential, null).getOrThrow().also {
                 it.shouldBeInstanceOf<CredentialIssuanceResult.Success>().also {