+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciPreAuthTest.kt
@@ -70,10 +70,10 @@ val OidvciPreAuthTest by testSuite {
     } - {
         test("process with pre-authorized code, credential offer, and authorization details for one credential") {
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
-            val credentialOffer = it.authorizationService.credentialOfferWithPreAuthnForUser(
+            val credentialOffer = it.authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = DummyUserProvider.user,
                 credentialIssuer = it.issuer.publicContext,
-                credentials = setOf(AtomicAttribute2023 to PLAIN_JWT)
+                schemes = setOf(AtomicAttribute2023 to PLAIN_JWT)
             )
             val credentialFormat = it.issuer.metadata.supportedCredentialConfigurations!![credentialIdToRequest]
                 .shouldNotBeNull()
@@ -102,10 +102,10 @@ val OidvciPreAuthTest by testSuite {
         }
 
         test("process with pre-authorized code, credential offer, and authorization details for all credentials") {
-            val credentialOffer = it.authorizationService.credentialOfferWithPreAuthnForUser(
+            val credentialOffer = it.authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = DummyUserProvider.user,
                 credentialIssuer = it.issuer.publicContext,
-                credentials = emptySet(),
+                schemes = emptySet(),
             )
             val credentialIdsToRequest = credentialOffer.configurationIds
                 .shouldHaveSize(4) // Atomic Attribute in 3 representations (JWT, ISO, dc+sd-jwt), mDL in ISO
@@ -140,10 +140,10 @@ val OidvciPreAuthTest by testSuite {
         }
 
         test("process with pre-authorized code, credential offer, and scope") {
-            val credentialOffer = it.authorizationService.credentialOfferWithPreAuthnForUser(
+            val credentialOffer = it.authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = DummyUserProvider.user,
                 credentialIssuer = it.issuer.publicContext,
-                credentials = emptySet(),
+                schemes = emptySet(),
             )
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
             // OID4VCI 5.1.2 Using scope Parameter to Request Issuance of a Credential
@@ -183,10 +183,10 @@ val OidvciPreAuthTest by testSuite {
         }
 
         test("two proofs over different keys lead to two credentials") {
-            val credentialOffer = it.authorizationService.credentialOfferWithPreAuthnForUser(
+            val credentialOffer = it.authorizationService.offerWithPreAuthnForUserForSchemes(
                 user = DummyUserProvider.user,
                 credentialIssuer = it.issuer.publicContext,
-                credentials = emptySet(),
+                schemes = emptySet(),
             )
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)